+++ src/ContaoNoUiSliderBundle.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace HeimrichHannot\NoUiSliderBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class ContaoNoUiSliderBundle extends Bundle{}