+++ src/HeimrichHannotNoUiSliderBundle.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace HeimrichHannot\NoUiSliderBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class HeimrichHannotNoUiSliderBundle extends Bundle
+{
+    public function getPath()
+    {
+        return \dirname(__DIR__);
+    }
+
+}